requirements.txt
Flask~=2.1.2
natsort
SQLAlchemy~=1.4.37
python-dotenv~=0.20.0
lagom~=1.7.1
requests~=2.27.1